requirements.txt
fastapi==0.116.1
pydantic==2.11.7
python_bcrypt==0.3.2
Requests==2.32.4
streamlit==1.47.1
